Fix helix package
We recently recived a bug report (helix-editor/helix#5090) from a user that was confused by the need to manually
symlink the runtime directory and asked us to improve our documentation in that regard.

We only document this requirement for manual builds because package manager usually provide a wrapper script that set s `HELIX_RUNTIME`.
I update the package to include such a script to avoid further problems. 


I also undid the removal of the `--frozen` flag. We track our `Cargo.lock` in git and keep it up to date so this should really not cause any problems.
I am not familiar with macports and don't have mac to test this with so feel free to edit this/use this a template.